+++ lib/src/json/sounds/_sounds.dart
@@ -1,2 +1,3 @@
+export 'audio_bus.dart';
 export 'custom_sound.dart';
 export 'sound.dart';